pyproject.toml
[project]
name = "websockets-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "langchain-openai>=0.3.33",
    "fastapi[standard]>=0.121.0",
]
